- Updated altar model - Adjusted dimension brightness and added a config entry - Added biome specific fog start and end - Added smooth fog transition from one biome to another

+++ java/thebetweenlands/world/biomes/BiomeSwampLands.java
@@ -9,7 +9,7 @@ public class BiomeSwampLands
 {
 	public BiomeSwampLands(int biomeID, BiomeDecoratorBaseBetweenlands decorator) {
 		super(biomeID, decorator);
-		this.setFogColor((short)10, (short)30, (short)12);
+		this.setFogColor((byte)10, (byte)30, (byte)12);
 		this.setHeightAndVariation(WorldProviderBetweenlands.LAYER_HEIGHT, 10);
 		this.setBiomeName("TestBiome");
 		this.setBlocks(BLBlockRegistry.betweenstone, BLBlockRegistry.swampDirt, BLBlockRegistry.swampGrass, BLBlockRegistry.swampDirt, BLBlockRegistry.betweenlandsBedrock);
